support withCredentials -- move tests + final touches
